+++ kaithem/src/astrallibwrapper.py
@@ -1,89 +1,98 @@
 # SPDX-FileCopyrightText: Copyright 2013 Daniel Dunn
 # SPDX-License-Identifier: GPL-3.0-only
 
-"""This file is an wrapper around some of the astral library, a pure python
+"""This file is an wrapper around some of the pyephem library, a pure python
 library providing all of kaithem's astronomical functions"""
 
-import calendar
 import datetime
-import time
 
-import astral.moon
-import astral.sun
-from astral import LocationInfo
+import ephem
 
 
-def getLocationInfo(lat, lon):
-    return LocationInfo("unknown", "unknown", "Etc/UTC", lat, lon)
+def getObserver(lat, lon, date):
+    o = ephem.Observer()
+    o.lat = str(lat)
+    o.lon = str(lon)
+    o.date = date
+    return o
 
 
-def dawn(lat, lon, date=None, elevation=0):
-    "Given a latitude and longitude, return civil dawn time for any given date object as a unix timestamp(default to today)"
+def toPyEphem(date: datetime.date):
+    return date.strftime("%Y/%m/%d %H:%M:%S")
+
+
+def sun_angle(lat, lon, date=None):
     if date is None:
-        date = datetime.datetime.now(datetime.UTC).date()
+        date = datetime.datetime.now(datetime.UTC)
 
-    return calendar.timegm(astral.sun.sun(astral.Observer(lat, lon, elevation), date)["dawn"].timetuple())
+    date = date.astimezone(datetime.UTC)
 
+    date = toPyEphem(date)
 
-def dusk(lat, lon, date=None, elevation=0):
-    "Given a latitude and longitude, return civil dusk time for any given date object as a unix timestamp(default to today)"
-    if date is None:
-        date = datetime.datetime.now(datetime.UTC).date()
+    o = getObserver(lat, lon, date)
 
-    return calendar.timegm(astral.sun.sun(astral.Observer(lat, lon, elevation), date)["dusk"].timetuple())
+    sun = ephem.Sun(o)
+    sun.compute(o)
+    sun_angle = float(sun.alt) * 57.2957795  # Convert Radians to degrees
 
-
-def sunrise(lat, lon, date=None, elevation=0):
-    "Given a latitude and longitude, return sunrise time for any given date object as a unix timestamp(default to today)"
-    if date is None:
-        date = datetime.datetime.now(datetime.UTC).date()
-
-    return calendar.timegm(astral.sun.sun(astral.Observer(lat, lon, elevation), date)["sunrise"].timetuple())
+    return sun_angle
 
 
-def sunset(lat, lon, date=None, elevation=0):
-    if date is None:
-        date = datetime.datetime.now(datetime.UTC).date()
-    return calendar.timegm(astral.sun.sun(astral.Observer(lat, lon, elevation), date)["sunset"].timetuple())
-
-
-def rahu(lat, lon, date=None, elevation=0):
-    "Given a latitude and longitude, return a tuple of the start and end timestamps of the given date's rahukalaam period"
-    if date is None:
-        date = datetime.datetime.now(datetime.UTC).date()
-    r = astral.rahukaalam(astral.Observer(lat, lon, elevation), date)
-    return (
-        calendar.timegm(r["start"].timetuple()),
-        calendar.timegm(r["end"].timetuple()),
-    )
+def is_sun_above(lat, lon, horizon, date=None):
+    return sun_angle(lat, lon, date) > horizon
 
 
 def is_night(lat, lon):
-    return not (sunrise(lat, lon) <= time.time() <= sunset(lat, lon))
+    return not is_sun_above(lat, lon, -18)
 
 
 def is_day(lat, lon):
-    return sunrise(lat, lon) <= time.time() <= sunset(lat, lon)
+    return is_sun_above(lat, lon, -18)
 
 
 def is_dark(lat, lon):
-    return not (dawn(lat, lon) <= time.time() <= dusk(lat, lon))
+    return not is_sun_above(lat, lon, -6)
 
 
 def is_light(lat, lon):
-    return dawn(lat, lon) <= time.time() <= dusk(lat, lon)
+    return not is_sun_above(lat, lon, -18)
 
 
-def isRahu(lat, lon):
-    return rahu(lat, lon)[0] <= time.time() <= rahu(lat, lon)[1]
-
-
-def moon():
+def moon_age():
     """
-    return 0 to 28 depending on current moon phase..
-            | 0  = New moon
-            | 7  = First quarter
-            | 14 = Full moon
-            | 21 = Last quarter
+    return days since new moon
     """
-    return astral.moon.phase(datetime.datetime.now(datetime.timezone.utc))
+
+    date = datetime.datetime.now(datetime.UTC)
+    date = ephem.Date(date)
+
+    nnm = ephem.next_new_moon(date)
+    pnm = ephem.previous_new_moon(date)
+
+    lunation = (date - pnm) / (nnm - pnm)
+
+    return 29.530588853 * lunation
+
+
+def moon_phase():
+    """Returns a floating-point number from 0-1. where 0=new, 0.5=full, 1=new"""
+
+    date = datetime.datetime.now(datetime.UTC)
+    date = ephem.Date(date)
+
+    nnm = ephem.next_new_moon(date)
+    pnm = ephem.previous_new_moon(date)
+
+    lunation = (date - pnm) / (nnm - pnm)
+
+    # Note that there is a ephem.Moon().phase() command, but this returns the
+    # percentage of the moon which is illuminated. This is not really what we want.
+
+    return lunation
+
+
+def moon_illumination():
+    o = getObserver(0, 0, datetime.datetime.now(datetime.UTC))
+    moon = ephem.Moon(o)
+
+    return moon.phase